+++ nflow-engine/src/main/java/io/nflow/engine/workflow/executor/StateVariableValueTooLongException.java
@@ -1,0 +1,10 @@
+package io.nflow.engine.workflow.executor;
+
+public class StateVariableValueTooLongException extends RuntimeException {
+
+  private static final long serialVersionUID = 1L;
+
+  public StateVariableValueTooLongException(String message) {
+    super(message);
+  }
+}