+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## 6.0.1-SNAPSHOT (future release)
 
 **Highlights**
+- `nflow-engine`
+  - Check that state variable value fits into the database column
 
 **Details**
 - `nflow-engine`
+  - Throw `StateVariableValueTooLongException` if a state variable value that does not fit into the database column is detected. Checked in `StateExecution.setVariable`, `StateExecution.addWorkflows`, `StateExecution.addChildWorkflows`, `WorkflowInstanceService.insertWorkflowInstance` and when creating a new instance via REST API. If the exception is thrown during state processing and not handled by the state implementation, nFlow engine will catch the exception and retry state processing after delay configured by property `nflow.executor.stateVariableValueTooLongRetryDelay.minutes` (default is 60).
   - Dependency updates:
     - jetty 9.4.24.v20191120
     - junit4 4.13-rc-1